pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "university-backend"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "Django==5.1.5",
    "djangorestframework==3.15.2",
    "gunicorn==23.0.0",
    "whitenoise==6.8.2",
]